Split app-state into pure (cljc) and browser-dependent (cljs) versions
This is part of a larger reorganization where the dispatcher is moving
to the front as the primary client api because it is synchronous and
thus much simpler to test than using aether. This is OK because we
really don't need an async api or message broadcast within the client
-- no components are subscribing individually to the aether. That gets
to messy. It is simpler to coordinate via app-state, which may add
some crud to the app-state but that's OK.