typeahead state constructor fns document and verify typeahead widget
Over time I added more and more knobs to the typeahead widget until it
was so configurable it was hard to ensure its configuration resulted
in a sane user experience just from looking at the code. To combat
this I am experimenting with constructor fns for each intended
use-case that provide validation and an api tailored to each
particular use-case. Current examples are when the typeahead is used
as a search field and when used as a value editor.